update test records
