--- sdk/communication/Azure.Communication.PhoneNumbers/tests/SessionRecords/PhoneNumberAdministrationClientLiveTests/GetAllSupportedCountriesEnUsLocaleAsync.json
@@ -1,45 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "https://chitchat-test.communication.azure.com/administration/phonenumbers/countries?locale=en-US\u0026api-version=2020-07-20-preview1",
-      "RequestMethod": "GET",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Date": "Fri, 19 Feb 2021 02:06:53 GMT",
-        "traceparent": "00-6e7bad3c7a361d43ac54c7f876089d74-2247626f5d86b641-00",
-        "User-Agent": [
-          "azsdk-net-Communication.PhoneNumbers/1.0.0-alpha.20210218.1",
-          "(.NET 5.0.3; Microsoft Windows 10.0.18363)"
-        ],
-        "x-ms-client-request-id": "30a249b25a094000dc678e62d175ece1",
-        "x-ms-content-sha256": "Sanitized",
-        "x-ms-return-client-request-id": "true"
-      },
-      "RequestBody": null,
-      "StatusCode": 200,
-      "ResponseHeaders": {
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Fri, 19 Feb 2021 02:06:54 GMT",
-        "MS-CV": "LWsbAgNK5E6InbRPhPjYdA.0",
-        "Request-Context": "appId=",
-        "Transfer-Encoding": "chunked",
-        "X-Azure-Ref": "0PB0vYAAAAACSYBu4ANo1TrUuORjEk2F8WVZSMzBFREdFMDMxMwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
-        "X-Processing-Time": "1563ms"
-      },
-      "ResponseBody": {
-        "countries": [
-          {
-            "localizedName": "United States",
-            "countryCode": "US"
-          }
-        ],
-        "nextLink": null
-      }
-    }
-  ],
-  "Variables": {
-    "COMMUNICATION_CONNECTION_STRING": "endpoint=https://chitchat-test.communication.azure.com/;accesskey=Kg==",
-    "RandomSeed": "815011431"
-  }
-}
