--- sdk/communication/Azure.Communication.PhoneNumbers/tests/SessionRecords/PhoneNumberAdministrationClientLiveTests/GetPhonePlans(%en-US%).json
@@ -1,110 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "https://chitchat-test.communication.azure.com/administration/phonenumbers/countries/US/phoneplangroups?locale=en-US\u0026api-version=2020-07-20-preview1",
-      "RequestMethod": "GET",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Date": "Fri, 19 Feb 2021 02:06:23 GMT",
-        "traceparent": "00-50ef618c115fe64d82ee8b5951a22eaf-07030722f6cec64a-00",
-        "User-Agent": [
-          "azsdk-net-Communication.PhoneNumbers/1.0.0-alpha.20210218.1",
-          "(.NET 5.0.3; Microsoft Windows 10.0.18363)"
-        ],
-        "x-ms-client-request-id": "616510a462ba67ae5073127e5b88837c",
-        "x-ms-content-sha256": "Sanitized",
-        "x-ms-return-client-request-id": "true"
-      },
-      "RequestBody": null,
-      "StatusCode": 200,
-      "ResponseHeaders": {
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Fri, 19 Feb 2021 02:06:22 GMT",
-        "MS-CV": "TcRgtL/VZk2i4dZpaJVn5g.0",
-        "Request-Context": "appId=",
-        "Transfer-Encoding": "chunked",
-        "X-Azure-Ref": "0Hh0vYAAAAABibzd\u002BWKwEQJijnOqvgrvxWVZSMzBFREdFMDMxMwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
-        "X-Processing-Time": "301ms"
-      },
-      "ResponseBody": {
-        "phonePlanGroups": [
-          {
-            "phonePlanGroupId": "55bc1415-9fe6-42d7-9ed4-5ea28c6a17cf",
-            "phoneNumberType": "Geographic",
-            "localizedName": "Azure- User - Geographic",
-            "localizedDescription": "These are numbers used by Azure resources."
-          },
-          {
-            "phonePlanGroupId": "671ee064-662f-4c3b-82a9-af2ab200dd5c",
-            "phoneNumberType": "Geographic",
-            "localizedName": "Azure - Geographic",
-            "localizedDescription": "These are numbers used by Azure resources."
-          }
-        ],
-        "nextLink": null
-      }
-    },
-    {
-      "RequestUri": "https://chitchat-test.communication.azure.com/administration/phonenumbers/countries/US/phoneplangroups/55bc1415-9fe6-42d7-9ed4-5ea28c6a17cf/phoneplans?locale=en-US\u0026api-version=2020-07-20-preview1",
-      "RequestMethod": "GET",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Date": "Fri, 19 Feb 2021 02:06:23 GMT",
-        "traceparent": "00-0d102ce68f274f459c5c3c8af1d327ce-cbebda8ea51b7d4e-00",
-        "User-Agent": [
-          "azsdk-net-Communication.PhoneNumbers/1.0.0-alpha.20210218.1",
-          "(.NET 5.0.3; Microsoft Windows 10.0.18363)"
-        ],
-        "x-ms-client-request-id": "1641a4177d67de1a4c3dd693d419fa40",
-        "x-ms-content-sha256": "Sanitized",
-        "x-ms-return-client-request-id": "true"
-      },
-      "RequestBody": null,
-      "StatusCode": 200,
-      "ResponseHeaders": {
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Fri, 19 Feb 2021 02:06:22 GMT",
-        "MS-CV": "y5eZx6vcw0OTXO0jXh2RnA.0",
-        "Request-Context": "appId=",
-        "Transfer-Encoding": "chunked",
-        "X-Azure-Ref": "0Hh0vYAAAAADTsEso5tc7RpDsfdyPOgU8WVZSMzBFREdFMDMxMwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
-        "X-Processing-Time": "310ms"
-      },
-      "ResponseBody": {
-        "phonePlans": [
-          {
-            "phonePlanId": "01432411-5169-4665-b13e-3fa56c10e1d1",
-            "localizedName": "Outbound Only PSTN For User - Geographic",
-            "locationType": "Selection",
-            "areaCodes": [],
-            "capabilities": [
-              "OutboundCalling",
-              "UserAssignment",
-              "Geographic"
-            ],
-            "maximumSearchSize": 20
-          },
-          {
-            "phonePlanId": "b528a997-03bb-446e-af98-3d99877cf0ba",
-            "localizedName": "Inbound Only PSTN For User - Geographic",
-            "locationType": "Selection",
-            "areaCodes": [],
-            "capabilities": [
-              "InboundCalling",
-              "UserAssignment",
-              "Geographic"
-            ],
-            "maximumSearchSize": 20
-          }
-        ],
-        "nextLink": null
-      }
-    }
-  ],
-  "Variables": {
-    "COMMUNICATION_CONNECTION_STRING": "endpoint=https://chitchat-test.communication.azure.com/;accesskey=Kg==",
-    "RandomSeed": "1863365519"
-  }
-}
