--- sdk/communication/Azure.Communication.PhoneNumbers/tests/SessionRecords/PhoneNumberAdministrationClientLiveTests/GetPhonePlans(%en-US%)Async.json
@@ -1,110 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "https://chitchat-test.communication.azure.com/administration/phonenumbers/countries/US/phoneplangroups?locale=en-US\u0026api-version=2020-07-20-preview1",
-      "RequestMethod": "GET",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Date": "Fri, 19 Feb 2021 02:06:59 GMT",
-        "traceparent": "00-a5a9841fd5499e4f8d71f504562df5b3-d4b3b52380d4c74c-00",
-        "User-Agent": [
-          "azsdk-net-Communication.PhoneNumbers/1.0.0-alpha.20210218.1",
-          "(.NET 5.0.3; Microsoft Windows 10.0.18363)"
-        ],
-        "x-ms-client-request-id": "59ce4b6a02979cbbff64c2c420cf4a3b",
-        "x-ms-content-sha256": "Sanitized",
-        "x-ms-return-client-request-id": "true"
-      },
-      "RequestBody": null,
-      "StatusCode": 200,
-      "ResponseHeaders": {
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Fri, 19 Feb 2021 02:06:58 GMT",
-        "MS-CV": "Qa1v5Y3u/0\u002BW36ibcjqE3g.0",
-        "Request-Context": "appId=",
-        "Transfer-Encoding": "chunked",
-        "X-Azure-Ref": "0Qh0vYAAAAACumPi08dNKS5W1nGiiy4jgWVZSMzBFREdFMDMxMwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
-        "X-Processing-Time": "448ms"
-      },
-      "ResponseBody": {
-        "phonePlanGroups": [
-          {
-            "phonePlanGroupId": "55bc1415-9fe6-42d7-9ed4-5ea28c6a17cf",
-            "phoneNumberType": "Geographic",
-            "localizedName": "Azure- User - Geographic",
-            "localizedDescription": "These are numbers used by Azure resources."
-          },
-          {
-            "phonePlanGroupId": "671ee064-662f-4c3b-82a9-af2ab200dd5c",
-            "phoneNumberType": "Geographic",
-            "localizedName": "Azure - Geographic",
-            "localizedDescription": "These are numbers used by Azure resources."
-          }
-        ],
-        "nextLink": null
-      }
-    },
-    {
-      "RequestUri": "https://chitchat-test.communication.azure.com/administration/phonenumbers/countries/US/phoneplangroups/55bc1415-9fe6-42d7-9ed4-5ea28c6a17cf/phoneplans?locale=en-US\u0026api-version=2020-07-20-preview1",
-      "RequestMethod": "GET",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Date": "Fri, 19 Feb 2021 02:06:59 GMT",
-        "traceparent": "00-a23f3ae418824546aba9f1d409c1726d-bd15a8cb13675244-00",
-        "User-Agent": [
-          "azsdk-net-Communication.PhoneNumbers/1.0.0-alpha.20210218.1",
-          "(.NET 5.0.3; Microsoft Windows 10.0.18363)"
-        ],
-        "x-ms-client-request-id": "d9daa191e8060dbdfaa37c32e101460c",
-        "x-ms-content-sha256": "Sanitized",
-        "x-ms-return-client-request-id": "true"
-      },
-      "RequestBody": null,
-      "StatusCode": 200,
-      "ResponseHeaders": {
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Fri, 19 Feb 2021 02:06:58 GMT",
-        "MS-CV": "ISkzyiED8UenF/7\u002BKHnG7g.0",
-        "Request-Context": "appId=",
-        "Transfer-Encoding": "chunked",
-        "X-Azure-Ref": "0Qh0vYAAAAAAeLMHdEtKoQqg92MGUEVKxWVZSMzBFREdFMDMxMwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
-        "X-Processing-Time": "326ms"
-      },
-      "ResponseBody": {
-        "phonePlans": [
-          {
-            "phonePlanId": "01432411-5169-4665-b13e-3fa56c10e1d1",
-            "localizedName": "Outbound Only PSTN For User - Geographic",
-            "locationType": "Selection",
-            "areaCodes": [],
-            "capabilities": [
-              "OutboundCalling",
-              "UserAssignment",
-              "Geographic"
-            ],
-            "maximumSearchSize": 20
-          },
-          {
-            "phonePlanId": "b528a997-03bb-446e-af98-3d99877cf0ba",
-            "localizedName": "Inbound Only PSTN For User - Geographic",
-            "locationType": "Selection",
-            "areaCodes": [],
-            "capabilities": [
-              "InboundCalling",
-              "UserAssignment",
-              "Geographic"
-            ],
-            "maximumSearchSize": 20
-          }
-        ],
-        "nextLink": null
-      }
-    }
-  ],
-  "Variables": {
-    "COMMUNICATION_CONNECTION_STRING": "endpoint=https://chitchat-test.communication.azure.com/;accesskey=Kg==",
-    "RandomSeed": "1728370529"
-  }
-}
