--- sdk/communication/Azure.Communication.PhoneNumbers/tests/SessionRecords/PhoneNumberAdministrationClientLiveTests/GetPhonePlans(null).json
@@ -1,110 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "https://chitchat-test.communication.azure.com/administration/phonenumbers/countries/US/phoneplangroups?api-version=2020-07-20-preview1",
-      "RequestMethod": "GET",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Date": "Fri, 19 Feb 2021 02:06:22 GMT",
-        "traceparent": "00-4ea193dc4268ba459f7bfafbfaa9c50b-b099dc685243a449-00",
-        "User-Agent": [
-          "azsdk-net-Communication.PhoneNumbers/1.0.0-alpha.20210218.1",
-          "(.NET 5.0.3; Microsoft Windows 10.0.18363)"
-        ],
-        "x-ms-client-request-id": "05696a5a50b365020f8c828c1a16c0f2",
-        "x-ms-content-sha256": "Sanitized",
-        "x-ms-return-client-request-id": "true"
-      },
-      "RequestBody": null,
-      "StatusCode": 200,
-      "ResponseHeaders": {
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Fri, 19 Feb 2021 02:06:21 GMT",
-        "MS-CV": "dbEOTvDJ7kGE\u002Bd2v1A588A.0",
-        "Request-Context": "appId=",
-        "Transfer-Encoding": "chunked",
-        "X-Azure-Ref": "0HR0vYAAAAADLzkq2v1aeTpoA38PxoHcTWVZSMzBFREdFMDMxMwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
-        "X-Processing-Time": "359ms"
-      },
-      "ResponseBody": {
-        "phonePlanGroups": [
-          {
-            "phonePlanGroupId": "55bc1415-9fe6-42d7-9ed4-5ea28c6a17cf",
-            "phoneNumberType": "Geographic",
-            "localizedName": "Azure- User - Geographic",
-            "localizedDescription": "These are numbers used by Azure resources."
-          },
-          {
-            "phonePlanGroupId": "671ee064-662f-4c3b-82a9-af2ab200dd5c",
-            "phoneNumberType": "Geographic",
-            "localizedName": "Azure - Geographic",
-            "localizedDescription": "These are numbers used by Azure resources."
-          }
-        ],
-        "nextLink": null
-      }
-    },
-    {
-      "RequestUri": "https://chitchat-test.communication.azure.com/administration/phonenumbers/countries/US/phoneplangroups/55bc1415-9fe6-42d7-9ed4-5ea28c6a17cf/phoneplans?api-version=2020-07-20-preview1",
-      "RequestMethod": "GET",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Date": "Fri, 19 Feb 2021 02:06:22 GMT",
-        "traceparent": "00-cb13a52db619bf4eb61304653ac5fd8c-9f09dd59d0ecbe46-00",
-        "User-Agent": [
-          "azsdk-net-Communication.PhoneNumbers/1.0.0-alpha.20210218.1",
-          "(.NET 5.0.3; Microsoft Windows 10.0.18363)"
-        ],
-        "x-ms-client-request-id": "7e31e22295b6387c7d21cd7520d2023d",
-        "x-ms-content-sha256": "Sanitized",
-        "x-ms-return-client-request-id": "true"
-      },
-      "RequestBody": null,
-      "StatusCode": 200,
-      "ResponseHeaders": {
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Fri, 19 Feb 2021 02:06:21 GMT",
-        "MS-CV": "e6HysrjfLkO4r\u002BX7pl2Ucg.0",
-        "Request-Context": "appId=",
-        "Transfer-Encoding": "chunked",
-        "X-Azure-Ref": "0Hh0vYAAAAAC5WyyJEoKbRr7fFl9BLbkKWVZSMzBFREdFMDMxMwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
-        "X-Processing-Time": "409ms"
-      },
-      "ResponseBody": {
-        "phonePlans": [
-          {
-            "phonePlanId": "01432411-5169-4665-b13e-3fa56c10e1d1",
-            "localizedName": "Outbound Only PSTN For User - Geographic",
-            "locationType": "Selection",
-            "areaCodes": [],
-            "capabilities": [
-              "OutboundCalling",
-              "UserAssignment",
-              "Geographic"
-            ],
-            "maximumSearchSize": 20
-          },
-          {
-            "phonePlanId": "b528a997-03bb-446e-af98-3d99877cf0ba",
-            "localizedName": "Inbound Only PSTN For User - Geographic",
-            "locationType": "Selection",
-            "areaCodes": [],
-            "capabilities": [
-              "InboundCalling",
-              "UserAssignment",
-              "Geographic"
-            ],
-            "maximumSearchSize": 20
-          }
-        ],
-        "nextLink": null
-      }
-    }
-  ],
-  "Variables": {
-    "COMMUNICATION_CONNECTION_STRING": "endpoint=https://chitchat-test.communication.azure.com/;accesskey=Kg==",
-    "RandomSeed": "1249678682"
-  }
-}
