--- sdk/communication/Azure.Communication.PhoneNumbers/tests/SessionRecords/PhoneNumberAdministrationClientLiveTests/GetPhonePlans(null)Async.json
@@ -1,110 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "https://chitchat-test.communication.azure.com/administration/phonenumbers/countries/US/phoneplangroups?api-version=2020-07-20-preview1",
-      "RequestMethod": "GET",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Date": "Fri, 19 Feb 2021 02:06:58 GMT",
-        "traceparent": "00-f67a5e5ddebd9043bc4f8cdabe59a37e-03ea6150e6d4064b-00",
-        "User-Agent": [
-          "azsdk-net-Communication.PhoneNumbers/1.0.0-alpha.20210218.1",
-          "(.NET 5.0.3; Microsoft Windows 10.0.18363)"
-        ],
-        "x-ms-client-request-id": "54079ccd269666fb3403e1ee00cc6a8e",
-        "x-ms-content-sha256": "Sanitized",
-        "x-ms-return-client-request-id": "true"
-      },
-      "RequestBody": null,
-      "StatusCode": 200,
-      "ResponseHeaders": {
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Fri, 19 Feb 2021 02:06:57 GMT",
-        "MS-CV": "hv\u002BThY6HhEmVCxoOCO4IoA.0",
-        "Request-Context": "appId=",
-        "Transfer-Encoding": "chunked",
-        "X-Azure-Ref": "0QR0vYAAAAAA140TBzAsjT7FjjaytnyfpWVZSMzBFREdFMDMxMwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
-        "X-Processing-Time": "301ms"
-      },
-      "ResponseBody": {
-        "phonePlanGroups": [
-          {
-            "phonePlanGroupId": "55bc1415-9fe6-42d7-9ed4-5ea28c6a17cf",
-            "phoneNumberType": "Geographic",
-            "localizedName": "Azure- User - Geographic",
-            "localizedDescription": "These are numbers used by Azure resources."
-          },
-          {
-            "phonePlanGroupId": "671ee064-662f-4c3b-82a9-af2ab200dd5c",
-            "phoneNumberType": "Geographic",
-            "localizedName": "Azure - Geographic",
-            "localizedDescription": "These are numbers used by Azure resources."
-          }
-        ],
-        "nextLink": null
-      }
-    },
-    {
-      "RequestUri": "https://chitchat-test.communication.azure.com/administration/phonenumbers/countries/US/phoneplangroups/55bc1415-9fe6-42d7-9ed4-5ea28c6a17cf/phoneplans?api-version=2020-07-20-preview1",
-      "RequestMethod": "GET",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Date": "Fri, 19 Feb 2021 02:06:58 GMT",
-        "traceparent": "00-b7ed5f5dd26c414193ff7bd59a1d5f5b-2e14c1d00f54fc4b-00",
-        "User-Agent": [
-          "azsdk-net-Communication.PhoneNumbers/1.0.0-alpha.20210218.1",
-          "(.NET 5.0.3; Microsoft Windows 10.0.18363)"
-        ],
-        "x-ms-client-request-id": "af2bb6f2ca2d7704be4f28f8125d6d0f",
-        "x-ms-content-sha256": "Sanitized",
-        "x-ms-return-client-request-id": "true"
-      },
-      "RequestBody": null,
-      "StatusCode": 200,
-      "ResponseHeaders": {
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Fri, 19 Feb 2021 02:06:57 GMT",
-        "MS-CV": "PMsoiY2pyEecOrdOJQ\u002BolA.0",
-        "Request-Context": "appId=",
-        "Transfer-Encoding": "chunked",
-        "X-Azure-Ref": "0Qh0vYAAAAABSloa/ehBMRrVjWzDN37PpWVZSMzBFREdFMDMxMwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
-        "X-Processing-Time": "295ms"
-      },
-      "ResponseBody": {
-        "phonePlans": [
-          {
-            "phonePlanId": "01432411-5169-4665-b13e-3fa56c10e1d1",
-            "localizedName": "Outbound Only PSTN For User - Geographic",
-            "locationType": "Selection",
-            "areaCodes": [],
-            "capabilities": [
-              "OutboundCalling",
-              "UserAssignment",
-              "Geographic"
-            ],
-            "maximumSearchSize": 20
-          },
-          {
-            "phonePlanId": "b528a997-03bb-446e-af98-3d99877cf0ba",
-            "localizedName": "Inbound Only PSTN For User - Geographic",
-            "locationType": "Selection",
-            "areaCodes": [],
-            "capabilities": [
-              "InboundCalling",
-              "UserAssignment",
-              "Geographic"
-            ],
-            "maximumSearchSize": 20
-          }
-        ],
-        "nextLink": null
-      }
-    }
-  ],
-  "Variables": {
-    "COMMUNICATION_CONNECTION_STRING": "endpoint=https://chitchat-test.communication.azure.com/;accesskey=Kg==",
-    "RandomSeed": "627743509"
-  }
-}
