--- sdk/communication/Azure.Communication.PhoneNumbers/tests/SessionRecords/PhoneNumberAdministrationClientLiveTests/GetPlanGroups(%en-US%,True).json
@@ -1,63 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "https://chitchat-test.communication.azure.com/administration/phonenumbers/countries/US/phoneplangroups?locale=en-US\u0026includeRateInformation=true\u0026api-version=2020-07-20-preview1",
-      "RequestMethod": "GET",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Date": "Fri, 19 Feb 2021 02:06:25 GMT",
-        "traceparent": "00-5235a922893d334482b2e351524d32ea-3f9ce7a6cb9ef347-00",
-        "User-Agent": [
-          "azsdk-net-Communication.PhoneNumbers/1.0.0-alpha.20210218.1",
-          "(.NET 5.0.3; Microsoft Windows 10.0.18363)"
-        ],
-        "x-ms-client-request-id": "6eeea0a97420258c8b12dbbd03e92622",
-        "x-ms-content-sha256": "Sanitized",
-        "x-ms-return-client-request-id": "true"
-      },
-      "RequestBody": null,
-      "StatusCode": 200,
-      "ResponseHeaders": {
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Fri, 19 Feb 2021 02:06:24 GMT",
-        "MS-CV": "RPDz8eiDi0WWvHMqfbat/g.0",
-        "Request-Context": "appId=",
-        "Transfer-Encoding": "chunked",
-        "X-Azure-Ref": "0IB0vYAAAAACBBTREKpqJQoPy0KkPOojeWVZSMzBFREdFMDMxMwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
-        "X-Processing-Time": "1400ms"
-      },
-      "ResponseBody": {
-        "phonePlanGroups": [
-          {
-            "phonePlanGroupId": "55bc1415-9fe6-42d7-9ed4-5ea28c6a17cf",
-            "phoneNumberType": "Geographic",
-            "localizedName": "Azure- User - Geographic",
-            "localizedDescription": "These are numbers used by Azure resources.",
-            "rateInformation": {
-              "monthlyRate": 1.0,
-              "currencyType": "USD",
-              "rateErrorMessage": null
-            }
-          },
-          {
-            "phonePlanGroupId": "671ee064-662f-4c3b-82a9-af2ab200dd5c",
-            "phoneNumberType": "Geographic",
-            "localizedName": "Azure - Geographic",
-            "localizedDescription": "These are numbers used by Azure resources.",
-            "rateInformation": {
-              "monthlyRate": 1.0,
-              "currencyType": "USD",
-              "rateErrorMessage": null
-            }
-          }
-        ],
-        "nextLink": null
-      }
-    }
-  ],
-  "Variables": {
-    "COMMUNICATION_CONNECTION_STRING": "endpoint=https://chitchat-test.communication.azure.com/;accesskey=Kg==",
-    "RandomSeed": "1954657508"
-  }
-}
