+++ sdk/communication/Azure.Communication.PhoneNumbers/tests/SessionRecords/PhoneNumbersClientLiveTests/ListPhoneNumbersAsync.json
@@ -1,16 +1,16 @@
 {
   "Entries": [
     {
-      "RequestUri": "https://sanitized.communication.azure.com/phoneNumbers?api-version=2021_03_07",
+      "RequestUri": "https://live-testing.communication.azure.com.127.0.0.1.xip.io:9998/phoneNumbers?api-version=2021_03_07",
       "RequestMethod": "GET",
       "RequestHeaders": {
         "Accept": "application/json",
         "Authorization": "Sanitized",
-        "Date": "Thu, 25 Feb 2021 03:21:20 GMT",
-        "traceparent": "00-9b07f014ab96094a913a7638da78e56a-b743508d91898240-00",
+        "Date": "Fri, 26 Feb 2021 06:10:37 GMT",
+        "traceparent": "00-f9a7d5de5b323e42a400fd1409f6271c-3f849ae5fcea0e42-00",
         "User-Agent": [
-          "azsdk-net-Communication.PhoneNumbers/1.0.0-alpha.20210224.1",
-          "(.NET Core 4.6.29719.03; Microsoft Windows 10.0.19042 )"
+          "azsdk-net-Communication.PhoneNumbers/1.0.0-alpha.20210225.1",
+          "(.NET 5.0.3; Microsoft Windows 10.0.19042)"
         ],
         "x-ms-client-request-id": "20ebe858182935f80f26182a75e5b17f",
         "x-ms-content-sha256": "Sanitized",
@@ -20,12 +20,12 @@
       "StatusCode": 200,
       "ResponseHeaders": {
         "Content-Type": "application/json; charset=utf-8",
-        "Date": "Thu, 25 Feb 2021 03:21:20 GMT",
-        "MS-CV": "69hTf/INC02NT0lVmap9\u002Bg.0",
+        "Date": "Fri, 26 Feb 2021 06:10:38 GMT",
+        "MS-CV": "rTUBNeJKuE2s2iqrPt3M4g.0",
         "Server": "Kestrel",
         "Strict-Transport-Security": "max-age=2592000",
         "Transfer-Encoding": "chunked",
-        "X-Processing-Time": "3ms"
+        "X-Processing-Time": "4ms"
       },
       "ResponseBody": {
         "phoneNumbers": [
@@ -119,6 +119,7 @@
     }
   ],
   "Variables": {
+    "COMMUNICATION_CONNECTION_STRING": "endpoint=https://live-testing.communication.azure.com.127.0.0.1.xip.io:9998/;accesskey=Kg==",
     "RandomSeed": "1246562785"
   }
 }
