+++ sdk/communication/Azure.Communication.PhoneNumbers/tests/SessionRecords/PhoneNumbersClientLiveTests/ListPhoneNumbersAsyncAsync.json
@@ -1,16 +1,16 @@
 {
   "Entries": [
     {
-      "RequestUri": "https://sanitized.communication.azure.com/phoneNumbers?api-version=2021_03_07",
+      "RequestUri": "https://live-testing.communication.azure.com.127.0.0.1.xip.io:9998/phoneNumbers?api-version=2021_03_07",
       "RequestMethod": "GET",
       "RequestHeaders": {
         "Accept": "application/json",
         "Authorization": "Sanitized",
-        "Date": "Thu, 25 Feb 2021 02:55:53 GMT",
-        "traceparent": "00-e96030f71a8ece418ab0f48a4dbfea3c-6b4bce434c3e894c-00",
+        "Date": "Fri, 26 Feb 2021 06:10:39 GMT",
+        "traceparent": "00-18194122267f524c8ba9260c3b4047c7-d7941dacbe93a548-00",
         "User-Agent": [
-          "azsdk-net-Communication.PhoneNumbers/1.0.0-alpha.20210224.1",
-          "(.NET Core 4.6.29719.03; Microsoft Windows 10.0.19042 )"
+          "azsdk-net-Communication.PhoneNumbers/1.0.0-alpha.20210225.1",
+          "(.NET 5.0.3; Microsoft Windows 10.0.19042)"
         ],
         "x-ms-client-request-id": "57914b5847fc57ab7d329f21e9d2526a",
         "x-ms-content-sha256": "Sanitized",
@@ -20,49 +20,15 @@
       "StatusCode": 200,
       "ResponseHeaders": {
         "Content-Type": "application/json; charset=utf-8",
-        "Date": "Thu, 25 Feb 2021 02:55:52 GMT",
-        "MS-CV": "l7xP1WzrSUi5h43diOdflQ.0",
+        "Date": "Fri, 26 Feb 2021 06:10:38 GMT",
+        "MS-CV": "J1TR30GyqUyogC6/yMVlFg.0",
         "Server": "Kestrel",
         "Strict-Transport-Security": "max-age=2592000",
         "Transfer-Encoding": "chunked",
-        "X-Processing-Time": "13ms"
+        "X-Processing-Time": "9ms"
       },
       "ResponseBody": {
         "phoneNumbers": [
-          {
-            "id": "Sanitized",
-            "phoneNumber": "Sanitized",
-            "countryCode": "US",
-            "phoneNumberType": "TollFree",
-            "capabilities": {
-              "calling": "none",
-              "sms": "outbound"
-            },
-            "assignmentType": "Application",
-            "purchaseDate": "0001-01-01T00:00:00\u002B00:00",
-            "cost": {
-              "amount": 0.0,
-              "currencyCode": "USD",
-              "billingFrequency": "Monthly"
-            }
-          },
-          {
-            "id": "Sanitized",
-            "phoneNumber": "Sanitized",
-            "countryCode": "US",
-            "phoneNumberType": "TollFree",
-            "capabilities": {
-              "calling": "none",
-              "sms": "outbound"
-            },
-            "assignmentType": "Application",
-            "purchaseDate": "0001-01-01T00:00:00\u002B00:00",
-            "cost": {
-              "amount": 0.0,
-              "currencyCode": "USD",
-              "billingFrequency": "Monthly"
-            }
-          },
           {
             "id": "Sanitized",
             "phoneNumber": "Sanitized",
@@ -153,6 +119,7 @@
     }
   ],
   "Variables": {
+    "COMMUNICATION_CONNECTION_STRING": "endpoint=https://live-testing.communication.azure.com.127.0.0.1.xip.io:9998/;accesskey=Kg==",
     "RandomSeed": "136581784"
   }
 }
