--- sdk/communication/Azure.Communication.PhoneNumbers/tests/SessionRecords/Sample2_PhoneNumberAdministrationClient/CreatePhoneNumberWithTokenCredentialAsync.json
@@ -1,38 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "https://chitchat-test.communication.azure.com/administration/phonenumbers/phonenumbers?api-version=2020-07-20-preview1",
-      "RequestMethod": "GET",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "traceparent": "00-e509e98acb7ece47b01f41c0dc5cf68f-46ff7dcb220b5743-00",
-        "User-Agent": [
-          "azsdk-net-Communication.PhoneNumbers/1.0.0-alpha.20210218.1",
-          "(.NET 5.0.3; Microsoft Windows 10.0.18363)"
-        ],
-        "x-ms-client-request-id": "c6b8b267b48a6e8c087dbfd82f1545fb",
-        "x-ms-return-client-request-id": "true"
-      },
-      "RequestBody": null,
-      "StatusCode": 200,
-      "ResponseHeaders": {
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Fri, 19 Feb 2021 02:05:47 GMT",
-        "MS-CV": "cVazD020QEuXgaD7jNI0wQ.0",
-        "Request-Context": "appId=",
-        "Transfer-Encoding": "chunked",
-        "X-Azure-Ref": "0/BwvYAAAAABIdS6Dc/vQRrNU1u8heg0zWVZSMzBFREdFMDMxMwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
-        "X-Processing-Time": "354ms"
-      },
-      "ResponseBody": {
-        "phoneNumbers": [],
-        "nextLink": null
-      }
-    }
-  ],
-  "Variables": {
-    "COMMUNICATION_CONNECTION_STRING": "endpoint=https://chitchat-test.communication.azure.com/;accesskey=Kg==",
-    "RandomSeed": "1875430357"
-  }
-}
