--- sdk/communication/Azure.Communication.PhoneNumbers/tests/SessionRecords/Sample2_PhoneNumberAdministrationClient/GetAllSupportedCountries.json
@@ -1,44 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "https://chitchat-test.communication.azure.com/administration/phonenumbers/countries?locale=en-US\u0026api-version=2020-07-20-preview1",
-      "RequestMethod": "GET",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Date": "Fri, 19 Feb 2021 02:05:46 GMT",
-        "User-Agent": [
-          "azsdk-net-Communication.PhoneNumbers/1.0.0-alpha.20210218.1",
-          "(.NET 5.0.3; Microsoft Windows 10.0.18363)"
-        ],
-        "x-ms-client-request-id": "035dd7e6d320cef0b9041c3c38a656a5",
-        "x-ms-content-sha256": "Sanitized",
-        "x-ms-return-client-request-id": "true"
-      },
-      "RequestBody": null,
-      "StatusCode": 200,
-      "ResponseHeaders": {
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Fri, 19 Feb 2021 02:05:46 GMT",
-        "MS-CV": "62thRzeJnUiGz9znhABexQ.0",
-        "Request-Context": "appId=",
-        "Transfer-Encoding": "chunked",
-        "X-Azure-Ref": "0\u002BRwvYAAAAAAsS4S9Q0AtQ7CmM32MVQ4PWVZSMzBFREdFMDMxMwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
-        "X-Processing-Time": "1636ms"
-      },
-      "ResponseBody": {
-        "countries": [
-          {
-            "localizedName": "United States",
-            "countryCode": "US"
-          }
-        ],
-        "nextLink": null
-      }
-    }
-  ],
-  "Variables": {
-    "COMMUNICATION_CONNECTION_STRING": "endpoint=https://chitchat-test.communication.azure.com/;accesskey=Kg==",
-    "RandomSeed": "662632677"
-  }
-}
