--- sdk/communication/Azure.Communication.PhoneNumbers/tests/SessionRecords/Sample2_PhoneNumberAdministrationClient/GetAllSupportedCountriesAsyncAsync.json
@@ -1,44 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "https://chitchat-test.communication.azure.com/administration/phonenumbers/countries?locale=en-US\u0026api-version=2020-07-20-preview1",
-      "RequestMethod": "GET",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Date": "Fri, 19 Feb 2021 02:05:50 GMT",
-        "User-Agent": [
-          "azsdk-net-Communication.PhoneNumbers/1.0.0-alpha.20210218.1",
-          "(.NET 5.0.3; Microsoft Windows 10.0.18363)"
-        ],
-        "x-ms-client-request-id": "28e9b728d3f2a2ee84f02e2e84f07370",
-        "x-ms-content-sha256": "Sanitized",
-        "x-ms-return-client-request-id": "true"
-      },
-      "RequestBody": null,
-      "StatusCode": 200,
-      "ResponseHeaders": {
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Fri, 19 Feb 2021 02:05:51 GMT",
-        "MS-CV": "ZGkiyn4340\u002BgLXRbypRfXQ.0",
-        "Request-Context": "appId=",
-        "Transfer-Encoding": "chunked",
-        "X-Azure-Ref": "0/RwvYAAAAAD2jmLVTgtiRI72V9O36AMVWVZSMzBFREdFMDMxMwA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
-        "X-Processing-Time": "1495ms"
-      },
-      "ResponseBody": {
-        "countries": [
-          {
-            "localizedName": "United States",
-            "countryCode": "US"
-          }
-        ],
-        "nextLink": null
-      }
-    }
-  ],
-  "Variables": {
-    "COMMUNICATION_CONNECTION_STRING": "endpoint=https://chitchat-test.communication.azure.com/;accesskey=Kg==",
-    "RandomSeed": "51555886"
-  }
-}
